[mojo-stdlib] Move code off using `Int/FloatLiteral` directly
This moves some code off using `Int`/`FloatLiteral` directly when they
actually want runtime types like `Int` or `Float32`.

LINALG_ORIG_REV_ID: 4c85e57a00afc441516814e7b3fdf28a70c74686

+++ mojo/kernels/src/linalg/fast_div.mojo
@@ -44,7 +44,7 @@ struct FastDiv[type: DType]:
         var cl = _ceillog2(UInt32(divisor))
         self._mprime = (
             (
-                (1 << bitwidthof[type]())
+                (UInt64(1) << bitwidthof[type]())
                 * ((1 << cl.cast[DType.uint64]()) - divisor)
                 / divisor
             )
